+++ src/gravity.hpp
@@ -14,9 +14,9 @@
 #include "vect.h"
 #include "body.hpp"
 
-const double G = 1.73205;
+const float	 G = 1.73205;
 const int N = 100000;
-const double OcttreeThreshold = 0.1;
+const float	 OcttreeThreshold = 0.1;
 
 extern int n;
 extern long long T;
